+++ packages/firebase_ml_vision/example/lib/picture_scanner.dart
@@ -31,6 +31,8 @@ class _PictureScannerState extends State<PictureScanner> {
   final TextRecognizer _recognizer = FirebaseVision.instance.textRecognizer();
   final TextRecognizer _cloudRecognizer =
       FirebaseVision.instance.cloudTextRecognizer();
+  final DocumentTextRecognizer _cloudDocumentRecognizer =
+      FirebaseVision.instance.cloudDocumentTextRecognizer();
 
   Future<void> _getAndScanImage() async {
     setState(() {
@@ -98,6 +100,9 @@ class _PictureScannerState extends State<PictureScanner> {
       case Detector.cloudText:
         results = await _cloudRecognizer.processImage(visionImage);
         break;
+      case Detector.cloudDocumentText:
+        results = await _cloudDocumentRecognizer.processImage(visionImage);
+        break;
       default:
         return;
     }
@@ -128,6 +133,9 @@ class _PictureScannerState extends State<PictureScanner> {
         break;
       case Detector.cloudText:
         painter = TextDetectorPainter(_imageSize, results);
+        break;
+      case Detector.cloudDocumentText:
+        painter = DocumentTextDetectorPainter(_imageSize, results);
         break;
       default:
         break;
@@ -196,6 +204,10 @@ class _PictureScannerState extends State<PictureScanner> {
               const PopupMenuItem<Detector>(
                 child: Text('Detect Cloud Text'),
                 value: Detector.cloudText,
+              ),
+              const PopupMenuItem<Detector>(
+                child: Text('Detect Document Text'),
+                value: Detector.cloudDocumentText,
               ),
             ],
           ),