+++ packages/firebase_ml_vision/example/test_driver/text_recognizer.dart
@@ -8,6 +8,10 @@ void textRecognizerTests() {
   group('$TextRecognizer', () {
     final TextRecognizer recognizer = FirebaseVision.instance.textRecognizer();
 
+    // TODO(bparrishMines): Unskip this test when text is recognized on ios.
+    // Related issues:
+    // https://github.com/FirebaseExtended/flutterfire/issues/1912
+    // https://github.com/FirebaseExtended/flutterfire/issues/2770
     test('processImage', () async {
       final String tmpFilename = await _loadImage('assets/test_text.png');
       final FirebaseVisionImage visionImage =
@@ -16,7 +20,7 @@ void textRecognizerTests() {
       final VisionText text = await recognizer.processImage(visionImage);
 
       expect(text.text, 'TEXT');
-    });
+    }, skip: (Platform.isIOS));
 
     test('close', () {
       expect(recognizer.close(), completes);